refactor: dissolve features/dialogs into shared/components
user-profile, user-menu, dialog-receiver and the avatar-selection
dialog are consumed by the shared header and by both the chat and
channel features, so they move to shared/components (the nested
avatar-selection/avatar-selection folder is flattened to
avatar-selection-dialog).
