+++ src/app/features/app_channel/components/add-member/add-member.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, inject, signal } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 
-import { DialogReceiverComponent } from '../../../dialogs/dialog-receiver/dialog-receiver.component';
+import { DialogReceiverComponent } from '../../../../shared/components/dialog-receiver/dialog-receiver.component';
 import { AuthService } from '../../../app_auth/services/auth/auth.service';
 import { User } from '../../../app_auth/models/user/user';
 import { ChannelService } from '../../services/channel/channel.service';
