+++ src/app/shared/components/avatar-selection-dialog/avatar-selection.component.html
