+++ src/app/shared/components/dialog-receiver/dialog-receiver.component.scss
