+++ src/app/shared/components/dialog-receiver/dialog-receiver.component.ts
@@ -1,8 +1,8 @@
 import { Component, inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
-import { NavigationService } from '../../../shared/services/navigation/navigation.service';
-import { User } from '../../app_auth/models/user/user';
+import { NavigationService } from '../../services/navigation/navigation.service';
+import { User } from '../../../features/app_auth/models/user/user';
 
 @Component({
   selector: 'app-dialog-receiver',
