+++ src/app/shared/components/user-menu/user-menu.component.scss
