+++ src/app/shared/components/user-profile/user-profile.component.scss
